+++ src/netspresso_trainer/pipelines/detection.py
@@ -33,6 +33,13 @@ class TwoStageDetectionPipeline(BasePipeline):
             model = model.to(device=devices)
             self.model = model
 
+        if conf.distributed:
+            self.backbone_to_train = model.module.backbone
+            self.head_to_train = model.module.head
+        else:
+            self.backbone_to_train = model.backbone
+            self.head_to_train = model.head
+
     def train_step(self, batch):
         self.model.train()
         images, labels, bboxes = batch['pixel_values'], batch['label'], batch['bbox']
@@ -43,8 +50,8 @@ class TwoStageDetectionPipeline(BasePipeline):
         self.optimizer.zero_grad()
 
         # forward to rpn
-        backbone = self.model.backbone
-        head = self.model.head
+        backbone = self.backbone_to_train
+        head = self.head_to_train
 
         features = backbone(images)['intermediate_features']
         if head.neck:
@@ -86,7 +93,7 @@ class TwoStageDetectionPipeline(BasePipeline):
         out = self.model(images)
 
         # Compute loss
-        head = self.model.head
+        head = self.head_to_train
         matched_idxs, roi_head_labels = head.roi_heads.assign_targets_to_proposals(out['boxes'], bboxes, labels)
         matched_gt_boxes = [bbox[idx] for idx, bbox in zip(matched_idxs, bboxes)]
         regression_targets = head.roi_heads.box_coder.encode(matched_gt_boxes, out['boxes'])